test(db): 添加并发测试以确保线程安全
- 为 getTableName 函数添加并发测试，验证其在多 goroutine 环境下的线程安全性
- 为 hasField 函数添加并发测试，确保字段缓存访问的线程安全
- 为 decode 函数添加并发测试，覆盖多种数据类型的并发解码场景
- 添加竞态条件测试，使用 -race 标志检测潜在的数据竞争问题
- 实现高并发压力测试，模拟大量并发操作下的系统稳定性
- 验证缓存有效性，确保并发环境下缓存的正确性和一致性
