--- middleware/storage/db/mock_client_test.go
@@ -1,90 +1,0 @@
-package db_test
-
-import (
-	"testing"
-
-	"github.com/lazygophers/lrpc/middleware/storage/db"
-	"gotest.tools/v3/assert"
-)
-
-// MinimalModel for testing models without standard fields
-type MinimalModel struct {
-	Name string
-}
-
-func (MinimalModel) TableName() string { return "minimal_models" }
-
-// TestModelWithTimestamps for testing models with timestamps
-type TestModelWithTimestamps struct {
-	ID   int64
-	Name string
-}
-
-func (TestModelWithTimestamps) TableName() string { return "test_models_with_timestamps" }
-
-// TestClientBasicMethods tests client methods that can be tested without actual DB
-func TestClientBasicMethods(t *testing.T) {
-	t.Run("test client driver type", func(t *testing.T) {
-		// Since the Client struct fields are not exported and we can't create
-		// a client without calling New(), skip this test for now
-		t.Skip("DriverType requires a properly initialized client")
-	})
-}
-
-// TestConfigApplyFunction tests the private apply function through client creation
-func TestConfigApplyFunction(t *testing.T) {
-	t.Run("test apply with different database types", func(t *testing.T) {
-		// Test various configurations to trigger the apply() function
-		configs := []*db.Config{
-			{Type: ""},           // Should default to sqlite
-			{Type: "sqlite3"},    // Should convert to sqlite
-			{Type: db.Sqlite},    // Should remain sqlite
-			{Type: db.MySQL},     // Should remain mysql
-			{Type: "postgres"},   // Should convert to postgres
-			{Type: "pg"},         // Should convert to postgres  
-			{Type: "postgresql"}, // Should convert to postgres
-			{Type: "pgsql"},      // Should convert to postgres
-			{Type: "sqlserver"},  // Should convert to sqlserver
-			{Type: "mssql"},      // Should convert to sqlserver
-		}
-		
-		for _, config := range configs {
-			// We can't easily test apply() directly since it's private
-			// and calling New() would require actual database setup
-			// For now, we can test that DSN() works with these configs
-			dsn := config.DSN()
-			assert.Assert(t, dsn != "" || dsn == "") // DSN might be empty for non-sqlite
-		}
-	})
-}
-
-// TestUtilityEdgeCases tests utility functions with edge cases
-func TestUtilityEdgeCases(t *testing.T) {
-	t.Run("test hasId, hasCreatedAt, hasUpdatedAt, hasDeletedAt functions", func(t *testing.T) {
-		// These functions are private but tested through NewModel
-		// Test model without standard fields
-		client := &db.Client{}
-		model := db.NewModel[MinimalModel](client)
-		assert.Assert(t, model != nil)
-		assert.Equal(t, "minimal_models", model.TableName())
-	})
-	
-	t.Run("test getTableName function", func(t *testing.T) {
-		// Test getTableName function through NewModel with different model types
-		
-		// Test model with standard fields
-		client := &db.Client{}
-		model := db.NewModel[TestModelWithTimestamps](client)
-		assert.Equal(t, "test_models_with_timestamps", model.TableName())
-	})
-}
-
-// TestSerializerInit tests that the JSON serializer is properly registered
-func TestSerializerInit(t *testing.T) {
-	t.Run("serializer registration", func(t *testing.T) {
-		// The init() function registers the serializer automatically
-		// We can test that creating a serializer works
-		serializer := &db.JsonSerializer{}
-		assert.Assert(t, serializer != nil)
-	})
-}
